projects/kbgen-suite/requirements.txt
crawl4ai>=0.4.0
rich>=13.7.0
typer>=0.12.3
pydantic>=2.7.0
PyYAML>=6.0.1
httpx>=0.27.0
qdrant-client>=1.9.0
sentence-transformers>=2.7.0
psycopg[binary]>=3.2.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
textual>=0.69.0
redis>=5.0.6
rq>=1.16.2
sse-starlette>=2.1.0
boto3>=1.34.0
numpy>=1.26.4
scikit-learn>=1.5.0
umap-learn>=0.5.6
hdbscan>=0.8.37
